feat!: Merge content script UI into a single API
