+++ src/client/content-scripts/ui/__tests__/index.test.ts
@@ -1,0 +1,15 @@
+import { describe } from 'node:test';
+
+describe('createContentScriptUi', () => {
+  describe('type=integrated', () => {});
+
+  describe('type=iframe', () => {});
+
+  describe('type=shadow-root', () => {});
+
+  describe('position=inline', () => {});
+
+  describe('position=overlay', () => {});
+
+  describe('position=modal', () => {});
+});
